Address nadav's code review, move JdbcProfileComponent into it's own file
